backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
timezonefinder>=6.4.0
astral==3.2
matplotlib==3.8.2
Pillow==10.0.1
requests==2.31.0
pytz==2023.3
geopy==2.4.1
reportlab==4.0.7
stripe==7.8.0
pywebpush==1.14.0
py-vapid>=1.9.1
httpx==0.25.2